+++ mne_icalabel/utils.py
@@ -1,9 +1,83 @@
-from typing import List, Tuple
-from numpy.typing import ArrayLike
+from typing import Tuple, List
+
+from mne.io import BaseRaw
+from numpy.typing import NDArray, ArrayLike
 import numpy as np
 
 
-def pol2cart(theta: ArrayLike, rho: ArrayLike) -> Tuple[ArrayLike, ArrayLike]:
+def mne_to_eeglab_locs(raw: BaseRaw) -> Tuple[NDArray[float], NDArray[float]]:
+    """Obtain EEGLab-like spherical coordinate from EEG channel positions.
+
+    TODO: @JACOB:
+    - Where is (0,0,0) defined in MNE vs EEGLab?
+    - some text description of how the sphere coordinates differ between MNE
+    and EEGLab.
+
+    Parameters
+    ----------
+    raw : mne.io.BaseRaw
+        Instance of raw object with a `mne.montage.DigMontage` set with
+        ``n_channels`` channel positions.
+
+    Returns
+    -------
+    Rd : np.array of shape (1, n_channels)
+        Angle in spherical coordinates of each EEG channel.
+    Th : np.array of shape (1, n_channels)
+        Degree in spherical coordinates of each EEG channel.
+    """
+
+    def _sph2topo(_theta, _phi):
+        """
+        Convert spherical coordinates to topo.
+        """
+        az = _phi
+        horiz = _theta
+        angle = -1 * horiz
+        radius = (np.pi / 2 - az) / np.pi
+        return angle, radius
+
+    def _cart2sph(_x, _y, _z):
+        """
+        Convert cartesian coordinates to spherical.
+        """
+        azimuth = np.arctan2(_y, _x)
+        elevation = np.arctan2(_z, np.sqrt(_x**2 + _y**2))
+        r = np.sqrt(_x**2 + _y**2 + _z**2)
+        # theta,phi,r
+        return azimuth, elevation, r
+
+    # get the channel position dictionary
+    montage = raw.get_montage()
+    positions = montage.get_positions()
+    ch_pos = positions["ch_pos"]
+
+    # get locations as a 2D array
+    locs = np.vstack(list(ch_pos.values()))
+
+    # Obtain carthesian coordinates
+    x = locs[:, 1]
+
+    # be mindful of the nose orientation in eeglab and mne
+    # TODO: @Jacob, please expand on this.
+    y = -1 * locs[:, 0]
+    # see https://github.com/mne-tools/mne-python/blob/24377ad3200b6099ed47576e9cf8b27578d571ef/mne/io/eeglab/eeglab.py#L105  # noqa
+    z = locs[:, 2]
+
+    # Obtain Spherical Coordinates
+    sph = np.array([_cart2sph(x[i], y[i], z[i]) for i in range(len(x))])
+    theta = sph[:, 0]
+    phi = sph[:, 1]
+
+    # Obtain Polar coordinates (as in eeglab)
+    topo = np.array([_sph2topo(theta[i], phi[i]) for i in range(len(theta))])
+    rd = topo[:, 1]
+    th = topo[:, 0]
+
+    return rd.reshape([1, -1]), np.degrees(th).reshape([1, -1])
+
+
+def pol2cart(theta: NDArray[float], rho: NDArray[float]) -> Tuple[NDArray[float], NDArray[float]]:
     """
     Converts polar coordinates to cartesian coordinates.
 
@@ -16,87 +90,7 @@ def pol2cart(theta: ArrayLike, rho: ArrayLike) -> Tuple[ArrayLike, ArrayLike]:
     return x, y
 
 
-def mergesimpts(
-    data: ArrayLike, tols: List[ArrayLike], mode: str = "average"
-) -> ArrayLike:
-    """
-
-    Args:
-        data (np.array): [description]
-        tols (list[np.array, np.array, np.array]): [description]
-        mode (str, optional): [description]. Defaults to 'average'.
-
-    Returns:
-        np.array: [description]
-    """
-    data_ = data.copy()[np.argsort(data[:, 0])]
-    newdata = []
-    tols_ = np.array(tols)
-    idxs_ready = []
-    point = 0
-    for point in range(data_.shape[0]):
-        if point in idxs_ready:
-            continue
-        else:
-            similar_pts = np.where(
-                np.prod(np.abs(data_ - data_[point]) < tols_, axis=-1)
-            )
-            similar_pts = np.array(list(set(similar_pts[0].tolist()) - set(idxs_ready)))
-            idxs_ready += similar_pts.tolist()
-            if mode == "average":
-                exemplar = np.mean(data_[similar_pts], axis=0)
-            else:
-                exemplar = data_[similar_pts].copy()[0]  # first
-            newdata.append(exemplar)
-    return np.array(newdata)
-
-
-def mergepoints2D(
-    x: ArrayLike, y: ArrayLike, v: ArrayLike
-) -> Tuple[ArrayLike, ArrayLike, ArrayLike]:
-    """
-    Averages values for points that are close to each other.
-
-    Args:
-        x (np.array): x-coordinates
-        y (np.array): y-coordinates
-        v (np.array): values
-
-    Returns:
-        tuple[np.array, np.array, np.array]: [description]
-    """
-    # Sort x and y so duplicate points can be averaged
-    # Need x,y and z to be column vectors
-    sz = x.size
-    x = x.copy()
-    y = y.copy()
-    v = v.copy()
-    x = np.reshape(x, sz, order="F")
-    y = np.reshape(y, sz, order="F")
-    v = np.reshape(v, sz, order="F")
-
-    myepsx = np.spacing(0.5 * (np.max(x) - np.min(x))) ** (1 / 3)
-    myepsy = np.spacing(0.5 * (np.max(y) - np.min(y))) ** (1 / 3)
-    # Look for x, y points that are identical (within a tolerance)
-    # Average out the values for these points
-    if np.all(np.isreal(v)):
-        data = np.stack((y, x, v), axis=-1)
-        yxv = mergesimpts(data, [myepsy, myepsx, np.inf], "average")
-        x = yxv[:, 1]
-        y = yxv[:, 0]
-        v = yxv[:, 2]
-    else:
-        # If z is imaginary split out the real and imaginary parts
-        data = np.stack((y, x, np.real(v), np.imag(v)), axis=-1)
-        yxv = mergesimpts(data, [myepsy, myepsx, np.inf, np.inf], "average")
-        x = yxv[:, 1]
-        y = yxv[:, 0]
-        # Re-combine the real and imaginary parts
-        v = yxv[:, 2] + 1j * yxv[:, 3]
-
-    return x, y, v
-
-
+# ----------------------------------------------------------------------------
 def gdatav4(
     x: ArrayLike, y: ArrayLike, v: ArrayLike, xq: ArrayLike, yq: ArrayLike
 ) -> Tuple[ArrayLike, ArrayLike, ArrayLike]:
@@ -145,3 +139,84 @@ def gdatav4(
                 0
             ]
     return xq, yq, vq
+
+
+def mergepoints2D(
+    x: ArrayLike, y: ArrayLike, v: ArrayLike
+) -> Tuple[ArrayLike, ArrayLike, ArrayLike]:
+    """
+    Averages values for points that are close to each other.
+
+    Args:
+        x (np.array): x-coordinates
+        y (np.array): y-coordinates
+        v (np.array): values
+
+    Returns:
+        tuple[np.array, np.array, np.array]: [description]
+    """
+    # Sort x and y so duplicate points can be averaged
+    # Need x,y and z to be column vectors
+    sz = x.size
+    x = x.copy()
+    y = y.copy()
+    v = v.copy()
+    x = np.reshape(x, sz, order="F")
+    y = np.reshape(y, sz, order="F")
+    v = np.reshape(v, sz, order="F")
+
+    myepsx = np.spacing(0.5 * (np.max(x) - np.min(x))) ** (1 / 3)
+    myepsy = np.spacing(0.5 * (np.max(y) - np.min(y))) ** (1 / 3)
+    # Look for x, y points that are identical (within a tolerance)
+    # Average out the values for these points
+    if np.all(np.isreal(v)):
+        data = np.stack((y, x, v), axis=-1)
+        yxv = mergesimpts(data, [myepsy, myepsx, np.inf], "average")
+        x = yxv[:, 1]
+        y = yxv[:, 0]
+        v = yxv[:, 2]
+    else:
+        # If z is imaginary split out the real and imaginary parts
+        data = np.stack((y, x, np.real(v), np.imag(v)), axis=-1)
+        yxv = mergesimpts(data, [myepsy, myepsx, np.inf, np.inf], "average")
+        x = yxv[:, 1]
+        y = yxv[:, 0]
+        # Re-combine the real and imaginary parts
+        v = yxv[:, 2] + 1j * yxv[:, 3]
+
+    return x, y, v
+
+
+def mergesimpts(
+    data: ArrayLike, tols: List[ArrayLike], mode: str = "average"
+) -> ArrayLike:
+    """
+
+    Args:
+        data (np.array): [description]
+        tols (list[np.array, np.array, np.array]): [description]
+        mode (str, optional): [description]. Defaults to 'average'.
+
+    Returns:
+        np.array: [description]
+    """
+    data_ = data.copy()[np.argsort(data[:, 0])]
+    newdata = []
+    tols_ = np.array(tols)
+    idxs_ready = []
+    point = 0
+    for point in range(data_.shape[0]):
+        if point in idxs_ready:
+            continue
+        else:
+            similar_pts = np.where(
+                np.prod(np.abs(data_ - data_[point]) < tols_, axis=-1)
+            )
+            similar_pts = np.array(list(set(similar_pts[0].tolist()) - set(idxs_ready)))
+            idxs_ready += similar_pts.tolist()
+            if mode == "average":
+                exemplar = np.mean(data_[similar_pts], axis=0)
+            else:
+                exemplar = data_[similar_pts].copy()[0]  # first
+            newdata.append(exemplar)
+    return np.array(newdata)
